requirements.txt
async-timeout==5.0.1
asyncpg==0.30.0
Authlib==1.5.0
azure-common==1.1.28
azure-core==1.33.0
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
bcrypt==4.3.0
boto3==1.37.1
botocore==1.37.1
fastapi==0.115.8
fastapi-cache2==0.2.2
httpcore==1.0.7
httpx==0.28.1
matplotlib==3.10.0
msal==1.32.0
msal-extensions==1.3.1
numpy==2.2.2
openai==1.61.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
sshtunnel==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
pytest==8.3.4
aiohttp==3.11.13
Authlib==1.5.0
oauthlib==3.2.2
async-lru==2.0.5
sentry-sdk==1.45.0